fixed base container class
